make elem.click() use random point
